# openapi.yaml
openapi: 3.0.3
info:
  title: Qubrax Commerce API
  description: Production-oriented E-commerce Backend API Contract
  version: 1.0.0
servers:
  - url: http://localhost:3000/api/v1
    description: Local development server
paths:
  /auth/register:
    post:
      tags:
        - Authentication
      summary: Register a new customer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterDto'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
  /auth/login:
    post:
      tags:
        - Authentication
      summary: Login to an existing account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginDto'
      responses:
        '201':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          description: Unauthorized
  
  /products:
    get:
      tags:
        - Catalog
      summary: List active products
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: cursor
          in: query
          schema:
            type: string
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Product'
                  meta:
                    $ref: '#/components/schemas/CursorPaginationMeta'

  /cart/items:
    post:
      tags:
        - Cart
      summary: Add an item to the cart
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - productId
                - quantity
              properties:
                productId:
                  type: string
                  format: uuid
                quantity:
                  type: integer
                  minimum: 1
      responses:
        '201':
          description: Item Added
          
  /checkout:
    post:
      tags:
        - Checkout
      summary: Finalize cart, reserve inventory, create order
      security:
        - bearerAuth: []
      parameters:
        - name: Idempotency-Key
          in: header
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CheckoutDto'
      responses:
        '201':
          description: Order Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '409':
          description: Conflict (e.g., Inventory unavailable, Duplicate Idempotency Key)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'

  /webhooks/payment:
    post:
      tags:
        - Webhooks
      summary: Handle asynchronous payment events from provider
      parameters:
        - name: stripe-signature
          in: header
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Event Received (Idempotent)

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    RegisterDto:
      type: object
      required:
        - email
        - password
        - firstName
        - lastName
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
          
    LoginDto:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
        password:
          type: string

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        refreshToken:
          type: string
          
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        sku:
          type: string
        name:
          type: string
        price:
          type: integer
          description: Price in minor units
        isActive:
          type: boolean
          
    CheckoutDto:
      type: object
      required:
        - shippingAddressId
      properties:
        shippingAddressId:
          type: string
          format: uuid
        couponCode:
          type: string
          
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        status:
          type: string
          enum: [PENDING, CONFIRMED, PAID, PROCESSING, SHIPPED, DELIVERED, CANCELLED]
        totalAmount:
          type: integer
          
    CursorPaginationMeta:
      type: object
      properties:
        nextCursor:
          type: string
          nullable: true
        hasMore:
          type: boolean

    ApiError:
      type: object
      properties:
        statusCode:
          type: integer
        code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        path:
          type: string